Adding headless tests for CASUALcore Reworking parts of CASUAL to run in headless mode Renaming PACKAGER to CASPACkager Upgrading to netbeans 7.4 for CASCADE, CASPACcreator, and CASPACkager Fixing remaining "compiling 1.6 source base under 1.7 compiler" errors. Adding catch-all create new file for zip.addToExistingZip Fixing CASPACcreator tests Fixing CASPACkager tests

+++ CASUALcore/src/CASUAL/Zip.java
@@ -157,6 +157,9 @@ public class Zip {
      * @throws IOException
      */
     public void addFilesToExistingZip(File[] files) throws IOException {
+        if (!outputZip.exists()){
+            outputZip.createNewFile();
+        }
         // get a temp file
         File tempFile = File.createTempFile(outputZip.getName(), null);
         // delete it, otherwise you cannot rename your existing zip to it.
@@ -555,7 +558,9 @@ public class Zip {
         //if rename fails, make copy
 
         if (!renameOk) {
-            tempFile.delete();
+            if (tempFile.exists()){
+                tempFile.delete();
+            }
             tempFile.createNewFile();
             FileOutputStream out;
             FileInputStream in = new FileInputStream(outputZip);
